truth and analytic refactoring.. maybe broke some things but overall this is in the right direction

+++ nars_logic/src/main/java/nars/bag/impl/CurveBag.java
@@ -358,7 +358,7 @@ public class CurveBag<K, E extends Item<K>> extends Bag<K, E> {
      * @return
      */
     public float getFocus(final float x) {
-        return (float) curve.y(x);
+        return curve.y(x);
     }
 
 //    public static long fastRound(final double d) {
